+++ R/zzz.R
@@ -40,7 +40,7 @@
 
 .onLoad <- function(libname = find.package("furniture"), pkgname = "furniture"){
   if(getRversion() >= "2.15.1") {
-    utils::globalVariables(c(".", ".rows"))
+    utils::globalVariables(c(".", ".rows", ":="))
   }
   invisible()
 }